[view refactor] - Adjustments to packages and the Table Controls Layout.  A couple of issues that remain are how to clear the Add Actor text fields.
